+++ pkg/front_end/lib/src/kernel/body_builder_helpers.dart
@@ -445,6 +445,7 @@ extension on MemberKind {
       case MemberKind.Catch:
       case MemberKind.Factory:
       case MemberKind.Local:
+      case MemberKind.AnonymousMethod:
       case MemberKind.NonStaticMethod:
       case MemberKind.StaticMethod:
       case MemberKind.TopLevelMethod: